refactor: childcomment, comment 패키지로 이동
Related to : #251
