+++ android/2023-emmsale/app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
             android:name=".presentation.ui.notificationBox.NotificationBoxActivity"
             android:exported="true" />
         <activity
-            android:name=".presentation.ui.childComments.ChildCommentsActivity"
+            android:name=".presentation.ui.eventdetail.comment.childComment.ChildCommentActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.ui.main.event.conferenceFilter.ConferenceFilterActivity"
