+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/eventdetail/EventDetailFragmentStateAdpater.kt
@@ -4,7 +4,7 @@ import EventParticipantFragment
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.emmsale.presentation.ui.comment.CommentsFragment
+import com.emmsale.presentation.ui.eventdetail.comment.CommentFragment
 import com.emmsale.presentation.ui.eventdetail.information.EventInfoFragment
 
 class EventDetailFragmentStateAdpater(
@@ -19,7 +19,7 @@ class EventDetailFragmentStateAdpater(
     override fun createFragment(position: Int): Fragment {
         return when (position) {
             INFORMATION_TAB -> EventInfoFragment.create(informationUrl, imageUrl)
-            COMMENT_TAB -> CommentsFragment.create(eventId)
+            COMMENT_TAB -> CommentFragment.create(eventId)
             PARTICIPANT_TAB -> EventParticipantFragment.create(eventId)
             else -> throw IllegalArgumentException("알수없는 ViewPager 오류입니다.")
         }
