+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/eventdetail/comment/uiState/CommentsUiState.kt
@@ -1,9 +1,9 @@
-package com.emmsale.presentation.ui.comment.uiState
+package com.emmsale.presentation.ui.eventdetail.comment.uiState
 
 import com.emmsale.data.comment.Comment
 import com.emmsale.data.member.Member1
 
-data class CommentsScreenUiState(
+data class CommentsUiState(
     val isNotLogin: Boolean = false,
     val isLoading: Boolean,
     val isError: Boolean,
@@ -11,7 +11,7 @@ data class CommentsScreenUiState(
     val comments: List<CommentUiState>,
 ) {
     companion object {
-        val Loading = CommentsScreenUiState(
+        val Loading = CommentsUiState(
             isLoading = true,
             isError = false,
             errorMessage = "",
@@ -21,7 +21,7 @@ data class CommentsScreenUiState(
         fun create(
             comments: List<Comment>,
             loginMember: Member1,
-        ) = CommentsScreenUiState(
+        ) = CommentsUiState(
             isLoading = false,
             isError = false,
             errorMessage = "",
