+++ android/2023-emmsale/app/src/main/res/layout/item_childcomments_childcomment.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="comment"
-            type="com.emmsale.presentation.ui.childComments.uiState.CommentUiState" />
+            type="com.emmsale.presentation.ui.eventdetail.comment.childComment.uiState.CommentUiState" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
